+++ neoforge/changelog.md
@@ -1,1 +1,2 @@
 - Fixed potential incompatibility with other mods
+- Fixed not being able to join servers without voice chat